+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/) and
 [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Changed
+- Minimum Ember version is now `3.16`.
+
 ## [1.0.0] - 2020-07-22
 
 - Upgraded package to Ember CLI `3.19`, requiring at least Node 10.