+++ CHANGELOG.md
@@ -8,6 +8,9 @@ project adheres to [Semantic Versioning](http://semver.org/) and
 
 ### Changed
 - **BREAKING:** Sweet Alert component no longer supports positional parameters.
+- **BREAKING:** The Sweet Alert component now follows the DDAU pattern. This means
+that the `onClose` action must be used in conjunction with the `show` attribute
+to ensure the show value is toggled between true/false.
 - Minimum Ember version is now `3.16`.
 
 ## [1.0.0] - 2020-07-22