+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/) and
 [this changelog format](http://keepachangelog.com/).
 
-## Unreleased
+## [2.0.0] - Unreleased
 
 ### Changed
 - **BREAKING:** Sweet Alert component no longer supports positional parameters.
@@ -13,8 +13,32 @@ that the `onClose` action must be used in conjunction with the `show` attribute
 to ensure the show value is toggled between true/false.
 - Minimum Ember version is now `3.16`.
 
+## [1.1.0] - Unreleased
+
+### Added
+- Updated list of configuration keys on the component to match those available
+in Sweet Alert. Added the following:
+  - `customContainerClass`
+  - `keydownListenerCapture`
+  - `cancelButtonColor`
+  - `showCloseButton`
+  - `validationMessage`
+  - `onAfterClose`
+- Update the list of methods available on the Sweet Alert service to match
+those available on Sweet Alert. The changes were:
+  - `getButtonWrapper` (added)
+  - `stopTimer` (added)
+  - `resumeTimer` (added)
+  - `toggleTimer` (added)
+  - `isTimerRunning` (added)
+  - `increaseTimer` (added)
+  - `showValidationMessage` (incorrectly listed as `showValidationError`)
+  - `resetValidationMessage` (incorrectly listed as `resetValidationError`)
+  - `getValidationMessage` (added)
+
 ## [1.0.0] - 2020-07-22
 
+### Changed
 - Upgraded package to Ember CLI `3.19`, requiring at least Node 10.
 - Minimum supported Ember version is now `3.8`.